Split Element components into several files
