+++ src/RhinoInside.Revit.GH/Components/Element/Delete.cs
@@ -1,0 +1,54 @@
+using System;
+using System.Linq;
+using Grasshopper.Kernel;
+
+namespace RhinoInside.Revit.GH.Components
+{
+  public class ElementDelete : TransactionsComponent
+  {
+    public override Guid ComponentGuid => new Guid("213C1F14-A827-40E2-957E-BA079ECCE700");
+    public override GH_Exposure Exposure => GH_Exposure.septenary | GH_Exposure.obscure;
+    protected override string IconTag => "X";
+
+    public ElementDelete()
+    : base("Element.Delete", "Delete", "Deletes elements from Revit document", "Revit", "Element")
+    { }
+
+    protected override void RegisterInputParams(GH_InputParamManager manager)
+    {
+      manager.AddParameter(new Parameters.Element(), "Elements", "E", "Elements to delete", GH_ParamAccess.tree);
+    }
+
+    protected override void RegisterOutputParams(GH_OutputParamManager manager) { }
+
+    protected override void TrySolveInstance(IGH_DataAccess DA)
+    {
+      if (!DA.GetDataTree<Types.Element>("Elements", out var elementsTree))
+        return;
+
+      var elementsToDelete = Parameters.Element.
+                             ToElementIds(elementsTree).
+                             GroupBy(x => x.Document).
+                             ToArray();
+
+      foreach (var group in elementsToDelete)
+      {
+        BeginTransaction(group.Key);
+
+        try
+        {
+          var deletedElements = group.Key.Delete(group.Select(x => x.Id).ToArray());
+
+          if (deletedElements.Count == 0)
+            AddRuntimeMessage(GH_RuntimeMessageLevel.Warning, "No elements were deleted");
+          else
+            AddRuntimeMessage(GH_RuntimeMessageLevel.Remark, $"{elementsToDelete.Length} elements and {deletedElements.Count - elementsToDelete.Length} dependant elements were deleted.");
+        }
+        catch (Autodesk.Revit.Exceptions.ArgumentException)
+        {
+          AddRuntimeMessage(GH_RuntimeMessageLevel.Error, "One or more of the elements cannot be deleted.");
+        }
+      }
+    }
+  }
+}
