+++ src/RhinoInside.Revit.GH/Components/Element/Materials.cs
@@ -1,0 +1,38 @@
+using System;
+using System.Linq;
+using Grasshopper.Kernel;
+using DB = Autodesk.Revit.DB;
+
+namespace RhinoInside.Revit.GH.Components
+{
+  public class ElementMaterials : Component
+  {
+    public override Guid ComponentGuid => new Guid("93C18DFD-FAAB-4CF1-A681-C11754C2495D");
+    protected override string IconTag => "M";
+
+    public ElementMaterials()
+    : base("Element.Materials", "Element.Materials", "Query element used materials", "Revit", "Element")
+    { }
+
+    protected override void RegisterInputParams(GH_InputParamManager manager)
+    {
+      manager.AddParameter(new Parameters.Element(), "Element", "E", "Element to query for its materials", GH_ParamAccess.item);
+    }
+
+    protected override void RegisterOutputParams(GH_OutputParamManager manager)
+    {
+      manager.AddParameter(new Parameters.Material(), "Materials", "M", "Materials this Element is made of", GH_ParamAccess.list);
+      manager.AddParameter(new Parameters.Material(), "Paint", "P", "Materials used to paint this Element", GH_ParamAccess.list);
+    }
+
+    protected override void TrySolveInstance(IGH_DataAccess DA)
+    {
+      DB.Element element = null;
+      if (!DA.GetData("Element", ref element))
+        return;
+
+      DA.SetDataList("Materials", element?.GetMaterialIds(false).Select(x => element.Document.GetElement(x)));
+      DA.SetDataList("Paint", element?.GetMaterialIds(true).Select(x => element.Document.GetElement(x)));
+    }
+  }
+}
