+++ src/RhinoInside.Revit.GH/Components/Element/Preview.cs
@@ -1,0 +1,61 @@
+using System;
+using System.Linq;
+using Grasshopper.Kernel;
+using Grasshopper.Kernel.Types;
+using DB = Autodesk.Revit.DB;
+
+namespace RhinoInside.Revit.GH.Components
+{
+  public class ElementPreview : Component
+  {
+    public override Guid ComponentGuid => new Guid("A95C7B73-6F70-46CA-85FC-A4402A3B6971");
+    protected override string IconTag => "P";
+
+    public ElementPreview()
+    : base("Element.Preview", "Preview", "Get the preview of the specified Element", "Revit", "Element")
+    { }
+
+    protected override void RegisterInputParams(GH_InputParamManager manager)
+    {
+      manager.AddParameter(new Parameters.Element(), "Element", "E", "Element to query", GH_ParamAccess.item);
+      manager[manager.AddParameter(new Parameters.Param_Enum<Types.ViewDetailLevel>(), "DetailLevel", "LOD", "Preview Level of detail [1, 3]", GH_ParamAccess.item)].Optional = true;
+      manager[manager.AddNumberParameter("Quality", "Q", "Meshes quality [0.0, 1.0]", GH_ParamAccess.item)].Optional = true;
+    }
+
+    protected override void RegisterOutputParams(GH_OutputParamManager manager)
+    {
+      manager.AddMeshParameter("Meshes", "M", "Element meshes", GH_ParamAccess.list);
+      manager.AddParameter(new Grasshopper.Kernel.Parameters.Param_OGLShader(), "Materials", "M", "Element materials", GH_ParamAccess.list);
+      manager.AddCurveParameter("Wires", "W", "Element wires", GH_ParamAccess.list);
+    }
+
+    protected override void TrySolveInstance(IGH_DataAccess DA)
+    {
+      var element = default(DB.Element);
+      if (!DA.GetData("Element", ref element))
+        return;
+
+      var detailLevel = DB.ViewDetailLevel.Undefined;
+      DA.GetData(1, ref detailLevel);
+      if (detailLevel == DB.ViewDetailLevel.Undefined)
+        detailLevel = DB.ViewDetailLevel.Coarse;
+
+      var relativeTolerance = double.NaN;
+      if (DA.GetData(2, ref relativeTolerance))
+      {
+        if (0.0 > relativeTolerance || relativeTolerance > 1.0)
+        {
+          AddRuntimeMessage(GH_RuntimeMessageLevel.Error, $"Parameter '{Params.Input[2].Name}' range is [0.0, 1.0].");
+          return;
+        }
+      }
+
+      var meshingParameters = !double.IsNaN(relativeTolerance) ? new Rhino.Geometry.MeshingParameters(relativeTolerance, Revit.VertexTolerance) : null;
+      Types.GeometricElement.BuildPreview(element, meshingParameters, detailLevel, out var materials, out var meshes, out var wires);
+
+      DA.SetDataList(0, meshes?.Select((x) => new GH_Mesh(x)));
+      DA.SetDataList(1, materials?.Select((x) => new GH_Material(x)));
+      DA.SetDataList(2, wires?.Select((x) => new GH_Curve(x)));
+    }
+  }
+}
